fix: quit the app when the window closes
Previously the electron process and several child processes would stick around after the main window was closed because `usbDetection.startMonitoring` keeps the process alive. So to ensure the app quits properly we need to stop monitoring for USB devices.

This commit ensures we stop monitoring by adding an optional cleanup for each IPC function handler registration. In the case of `getDeviceList` and `manageDeviceSubscription` we assert that we're using USB monitoring, but then release that assertion during cleanup. The cleanup functions are called when the main browser window is closed.