refactor(usb): reduce the complexity of USB observers
I had a hard time following the logic of the rxjs operators after not having looked at it for a while, so I stopped using most of them and just did it "manually". Additionally, this moves the responsibility of creating the shared devices observer to the main `index.ts` file.