fix: unmount anything missing its source
When a drive is removed but not unmounted it will still have a mount entry, but its source will be missing. This changes the auto-cleanup code in `getUsbDrives` to look at all mounted file systems to determine which if any are missing their source. Those which are will be unmounted.

This means that we're no longer restricted to cleaning up mounts matching `/**/usb-drive*`.