fix: set datetime & timezone correctly
Setting the time first and then the timezone results in the wrong time being set when the timezone differs from the current system timezone. This is because setting the time to e.g. 12:00:00 when in Pacific Time and then setting the time zone to Eastern Time will update the time to 15:00:00.

Closes votingworks/vxsuite#410