chore(electron): upgrade Electron to v17
In the process, it was easier to switch from the seemingly not really maintained `electron-forge` to using `electron-builder` and `electron` directly. This did introduce one wrinkle: `electron-builder` uses `fpm` to build `.deb` packages, which itself depends on Ruby. `electron-builder` tries to use a self-contained `fpm` and Ruby build, but it doesn't work on ARM yet (https://github.com/electron-userland/electron-builder/issues/5154). I work around this problem by telling `electron-build` (via an environment variable) to use the system version of `fpm`, then ensure that Ruby and the `fpm` gem are installed in `install-dependencies.sh`.