chore(dev-deps): upgrade TS/jest/eslint/prettier
Upgrading prettier changed a bunch of the code formatting, specificially changing `a => a` to `(a) => a`. I also took the opportunity to switch to `semi: true` to match vxsuite, so there's a lot of diff noise from that.

`@typescript-eslint` updated their rules a bunch, including one that makes it an error to assign an `any` value, hence all the explicit type coercion (i.e. `as Foo`). Also more strict use of `Promise`-returning functions, so there's more `await …` or `void …` to explicitly handle or not handle the `Promise` return value.