test(integration): actually test a few things in Electron
There are tools that are better suited to running tests in Electron, but they're all kinda heavy (see https://webdriver.io). So instead, I abuse `vite` for fun and profit! Uses a custom vite plugin to create a test server for both development and CI. It's actually kinda great: we can use TypeScript and automatically rebuild/retest on file save in development, and in CI it runs once and exits.