fix: make `kiosk.devices` & `kiosk.printers` work again
Fixes #120

Replaces the RxJS `Observable` with a very lightweight `KioskBrowser.Observable` that only has `subscribe` with a function (a la RxJS `next`), no completion or error callbacks since we don't really need those. We're still using RxJS under the hood, but I've come to regret that. We should consider replacing it, but this is not the time.

Also uses `KioskBrowser.Device` rather than the one from `usb-detection`. They should be the same, but using the former means we can avoid unnecessary imports of `usb-detection`.